fix(layout): split humongous components into subs and move things around a bit
