+++ src/pages/events.astro
@@ -1,7 +1,7 @@
 ---
 import Layout from '../layouts/Layout.astro';
-import Event from '../components/Events.svelte';
-import Header from '../components/Header.svelte';
+import Event from '../components/app/pages/Event.svelte';
+import Header from '../components/app/Header.svelte';
 import { lucia } from '../lib/lucia';
 import { DUMMY_SESSION } from '../lib/helpers';
