+++ src/pages/events/[uuid].astro
@@ -1,7 +1,7 @@
 ---
 import Layout from '../../layouts/Layout.astro';
-import Challenge from '../../components/Challenges.svelte';
-import Header from '../../components/Header.svelte';
+import Challenge from '../../components/app/pages/Challenges.svelte';
+import Header from '../../components/app/Header.svelte';
 import DatabaseActions from '../../lib/actions';
 import { lucia } from '../../lib/lucia';
 import { DUMMY_SESSION } from '../../lib/helpers';
