+++ src/pages/index.astro
@@ -1,7 +1,7 @@
 ---
 import Layout from '../layouts/Layout.astro';
-import Dashboard from '../components/Dashboard.svelte';
-import Header from '../components/Header.svelte';
+import Dashboard from '../components/app/pages/Dashboard.svelte';
+import Header from '../components/app/Header.svelte';
 import { lucia } from '../lib/lucia';
 import { DUMMY_SESSION } from '../lib/helpers';
