+++ src/pages/reset/password.astro
@@ -1,7 +1,7 @@
 ---
 import Layout from '../../layouts/Layout.astro';
-import PasswordReset from '../../components/PasswordResetToken.svelte';
-import Header from '../../components/Header.svelte';
+import PasswordReset from '../../components/app/pages/PasswordToken.svelte';
+import Header from '../../components/app/Header.svelte';
 ---
 
 <Layout title="Password Reset">
