+++ src/pages/scoreboards.astro
@@ -1,7 +1,7 @@
 ---
 import Layout from '../layouts/Layout.astro';
-import Scoreboard from '../components/Scoreboards.svelte';
-import Header from '../components/Header.svelte';
+import Scoreboard from '../components/app/pages/Scoreboard.svelte';
+import Header from '../components/app/Header.svelte';
 import { lucia } from '../lib/lucia';
 import { DUMMY_SESSION } from '../lib/helpers';
