+++ src/pages/signup.astro
@@ -1,7 +1,7 @@
 ---
 import Layout from '../layouts/Layout.astro';
-import SignupPage from '../components/SignUp.svelte';
-import Header from '../components/Header.svelte';
+import SignupPage from '../components/app/pages/Signup.svelte';
+import Header from '../components/app/Header.svelte';
 import { lucia } from '../lib/lucia';
 
 if (Astro.locals.user) {
