+++ src/pages/teams.astro
@@ -1,7 +1,7 @@
 ---
 import Layout from '../layouts/Layout.astro';
-import TeamOverview from '../components/TeamOverview.svelte';
-import Header from '../components/Header.svelte';
+import TeamOverview from '../components/app/pages/Team.svelte';
+import Header from '../components/app/Header.svelte';
 import { lucia } from '../lib/lucia';
 import { DUMMY_SESSION } from '../lib/helpers';
